# Simplify aegraphs by removing union-find and canonical eclass IDs.