--- a/cranelift/codegen/src/ctxhash.rs
+++ b/cranelift/codegen/src/ctxhash.rs
@@ -114,6 +114,69 @@ impl<K, V> CtxHashMap<K, V> {
             })
             .map(|bucket| &bucket.v)
     }
+
+    /// Look up a key, returning an `Entry` that refers to an existing
+    /// value or allows inserting a new one.
+    pub fn entry<'a, Ctx>(&'a mut self, k: K, ctx: &Ctx) -> Entry<'a, K, V>
+    where
+        Ctx: CtxEq<K, K> + CtxHash<K>,
+    {
+        let hash = compute_hash(ctx, &k);
+        let raw = self.raw.entry(
+            hash as u64,
+            |bucket| hash == bucket.hash && ctx.ctx_eq(&bucket.k, &k),
+            |bucket| compute_hash(ctx, &bucket.k) as u64,
+        );
+        match raw {
+            hashbrown::hash_table::Entry::Occupied(o) => Entry::Occupied(OccupiedEntry { raw: o }),
+            hashbrown::hash_table::Entry::Vacant(v) => Entry::Vacant(VacantEntry {
+                hash,
+                key: k,
+                raw: v,
+            }),
+        }
+    }
+}
+
+/// A reference to an existing or vacant entry in the hash table.
+pub enum Entry<'a, K, V> {
+    Occupied(OccupiedEntry<'a, K, V>),
+    Vacant(VacantEntry<'a, K, V>),
+}
+
+/// A reference to an occupied entry in the hash table.
+pub struct OccupiedEntry<'a, K, V> {
+    raw: hashbrown::hash_table::OccupiedEntry<'a, BucketData<K, V>>,
+}
+
+/// A reference to a vacant entry in the hash table.
+pub struct VacantEntry<'a, K, V> {
+    hash: u32,
+    key: K,
+    raw: hashbrown::hash_table::VacantEntry<'a, BucketData<K, V>>,
+}
+
+impl<'a, K, V> OccupiedEntry<'a, K, V> {
+    /// Get the existing value.
+    pub fn get(&self) -> &V {
+        &self.raw.get().v
+    }
+
+    /// Get the existing value, mutably.
+    pub fn get_mut(&mut self) -> &mut V {
+        &mut self.raw.get_mut().v
+    }
+}
+
+impl<'a, K, V> VacantEntry<'a, K, V> {
+    /// Insert a new value.
+    pub fn insert(self, v: V) {
+        self.raw.insert(BucketData {
+            hash: self.hash,
+            k: self.key,
+            v,
+        });
+    }
 }
 
 #[cfg(test)]