@alexcrichton 2025-04-14
Looking through some open fuzz bugs this morning we've got two timeouts which I think may want to be investigated as a result of this. Typically I just ignore timeouts since we expect those to happen occasionally, but both timeouts show the "regression range" as https://github.com/bytecodealliance/wasmtime/compare/eafe743901f20eab0c4da3ffaa4fac3783e51be8...7e5a7e521ae121e239a8ad9e927653b26e474ac0 which notably has this PR as the main change. Additionally both timeouts were in the egraph passes which we typically don't time out in. Given all that, @cfallin would you be up for digging into these and seeing if the timeout is relevant? I think it's reasonable if the conclusion is "yep that's expected" but I think it'd be good to confirm ideally.

* https://oss-fuzz.com/testcase-detail/4806924172591104
* https://oss-fuzz.com/testcase-detail/5226667534778368